Update for AuthManager, somewhat
Since WMF only uses the extension on wikitech.wikimedia.org and is going
to stop using it there once OpenStackManager is replaced, and I don't
know much about LDAP, the update here is mainly adapting the core BC
AuthPluginPrimaryAuthenticationProvider to slightly-more-specifically
wrap LdapAuthenticationPlugin to get things working well enough for
wikitech.

If someone who knows about authenticating against LDAP wants to pick up
the torch and write a really good PrimaryAuthenticationProvider that
doesn't have craziness like a global "domain" variable that we have to
guess at the correct value in half the entry points, I'll be happy to
help you with that.

Bug: T110453
Change-Id: If44ca97dc72dcad9e7935eced805ff0f1f12c093