when user lookups fail, NULL out $this->userInfo
Since we use this class as a singleton object,
a failed user lookup could result in us 'returning'
incorrect data in $this->userInfo based on whatever
was the last successful lookup.

Probably the caller is meant to check $this->fetchedUserInfo
before checking $this->userInfo but a) fetchedUserInfo
also wasn't unset in all cases and b) OpenStackManager
definitely did not check that.

Bug: T171313
Change-Id: Ia2e2df5c09392e00e458fbe41c9782e208479498