Expose methods used by OpenStackManager
The introduction of public and private keywords in 9f73202ae3c7 made
four methods used by the OpenStackManager extension private. This change
sets those methods as public. This is not an endorsement of the levels
of interesting encapsulation abuse that OpenStackManager relies on, but
it is nicer than breaking wikitech.wikimedia.org and faster than
attempting to remove the dependencies between OpenStackManager and
LdapAuthentication.

Bug: T208995
Change-Id: If5cc2b4953e05177ebf0165a60a4a58349e92123