Handle autocreation via LdapAuthenticationPlugin::initUser
Create a record in the backing directory when
LdapAuthenticationPlugin::initUser is called with the $autocreate=true
argument. A random password will be assigned upon creation with the
assumption that the caller is either going to immediately set a known
password (as would happen with createAndPromote.php) or that the
intended user will have other means of password recovery.

Bug: T212689
Change-Id: Iebb9c648adcdabe7c4878300f6829f99406c2174