Also set an LDAP password policy on Block
When a user gets blocked and per the findings in T168692, we also set a
configurable password policy. The exact parameters of the password
policy should be configured on the LDAP server. Absence of it would lead
to users not being locked out, so care must be taken against that (there
is no way around that).

Per T168692 and  in order to minimize the chances of a user being
unlocked by mistake, the configured password policy on the LDAP server
should enforce a max age of a password of 1 sec. That closes the
loophole of the pwdAccountLockedTime attribute being set to
000001010000Z and the password then being updated via some other method
than mediawiki.

Bug: T168692
Change-Id: Ib40934155ae040fc48678415d39059d476be871d