Discard LdapAuthenticationPlugin->ldapconn after unbind
PHP's ldap_unbind() marks the ldap connection resource as unusable.
There is no easy method to determine if a given LDAP resource is valid
or not, so clearing the member variable after unbind is the easiest way
to ensure that calling code can determine if there is a usable
connection to the LDAP directory or not.

Bug: T240734
Change-Id: I7001dd529f0194a52a505b260ba9a128a71a6778