Provide i18n for old AuthPlugin messages
This extension relies on core providing these messages which
were used by the old AuthPlugin system. The AuthPlugin code got
removed entirely in I330b30d and this extension was not updated
for the messages.

Using the messages keys and values as they were since none is
likely to be used by core again and given that the extension
itself is only used on wikitech and will be removed from there
in the not too distant future.

Bug: T298508
Change-Id: I568d4989cb3720e846a082f4a936e5377f5f9867